requirements.txt
libsvm
matplotlib~=3.7.2
pytorch
seaborn
sklearn
tqdm
numpy~=1.25.2
opencv-python
opencv-python-headless
pandas
pathlib
pillow~=9.5.0
torch~=2.0.1
torchvision